requirements.txt
# Core FastAPI App
fastapi==0.110.0
uvicorn==0.29.0
requests==2.31.0
pydantic==2.6.4
python-dotenv==1.0.1
python-multipart==0.0.9

# Agent Infrastructure
openai==2.2.0
ollama==0.1.0
asyncpg==0.29.0
aiofiles==23.2.1
numpy==1.26.4
psutil==5.9.8

# Testing
pytest==8.2.0
httpx==0.25.2
pytest-asyncio==0.23.3

# Development Dashboard
flask==3.0.0
qdrant-client>=1.7.0
